pyproject.toml
[project]
name = "cigarbox-ai"
version = "0.1.0"
description = "AI research project for CigarBox solutions"
dependencies = [
    "numpy",
    "pandas",
    "requests",
    "matplotlib",
    "seaborn",
    "google-genai>=1.25.0",
    "openpyxl>=3.1.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.8.5",
    "markdown>=3.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "jupyter",
    "ipython",
]
ml = [
    "scikit-learn",
    "torch",
    "transformers",
    "datasets",
]
